Revert "Merge pull request #27163 from jrose-apple/laserwriter"
This reverts commit 258f372ee7b71ab7d459a9598486fc99f28dd86d, reversing
changes made to e675c4947d155c548b61b2ef27a255a4d436e37c.
